requirements-dev.txt
Sphinx==1.3.5
sphinx-rtd-theme==0.1.9
pytest
pytest-cov
pytest-django
